Create equilibrium
The function accepts an integer array 'arr' of size 'n' as its argument. The function needs to return the index of an equilibrium point in the array, where the sum of elements on the left of the index is equal to the sum of elements on the right of the index. If no equilibrium point exists return -1.

Input 1:
7  
-7 1 5 2 -4 3 0

Output:
3